Fix ESLint warnings and improve code structure
- Fixed Fast refresh warnings by separating constants from components
- Created separate metadata.ts and viewport.ts files for better organization
- Updated ESLint configuration to handle Next.js patterns properly
- Added eslint-disable comments for legitimate Next.js exports
- All tests passing and build successful
