+++ app/about/page.tsx
@@ -1,8 +1,9 @@
 import { ArrowRight } from 'lucide-react';
 import Link from 'next/link';
-import { aboutMetadata } from './constants';
+import { aboutMetadata } from './metadata';
 
 export const dynamic = 'force-static';
+// eslint-disable-next-line react-refresh/only-export-components
 export const metadata = aboutMetadata;
 
 export default function AboutPage() {
